Enums moved to separate namespace.
