+++ packages/graphql/src/SortBy/Contracts/Sorter.php
@@ -3,8 +3,8 @@
 namespace LastDragon_ru\LaraASP\GraphQL\SortBy\Contracts;
 
 use LastDragon_ru\LaraASP\GraphQL\Builder\Property;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Direction;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Nulls;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Enums\Direction;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Enums\Nulls;
 
 /**
  * @template TBuilder of object
