+++ packages/graphql/src/SortBy/Enums/Direction.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types = 1);
 
-namespace LastDragon_ru\LaraASP\GraphQL\SortBy;
+namespace LastDragon_ru\LaraASP\GraphQL\SortBy\Enums;
 
 use GraphQL\Type\Definition\Deprecated;
 use GraphQL\Type\Definition\Description;
