+++ packages/graphql/src/SortBy/Enums/Nulls.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types = 1);
 
-namespace LastDragon_ru\LaraASP\GraphQL\SortBy;
+namespace LastDragon_ru\LaraASP\GraphQL\SortBy\Enums;
 
 enum Nulls {
     case First;
