+++ packages/graphql/src/SortBy/Operators/FieldTest.php
@@ -8,8 +8,8 @@ use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use Laravel\Scout\Builder as ScoutBuilder;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Property;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Direction;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Directives\Directive;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Enums\Direction;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\EloquentSorter;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\QuerySorter;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters\ScoutSorter;
