+++ packages/graphql/src/SortBy/Sorters/QuerySorter.php
@@ -5,8 +5,8 @@ namespace LastDragon_ru\LaraASP\GraphQL\SortBy\Sorters;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Property;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Contracts\Sorter;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Direction;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Nulls;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Enums\Direction;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Enums\Nulls;
 use Override;
 
 use function implode;
