+++ packages/graphql/src/SortBy/Sorters/ScoutSorter.php
@@ -7,8 +7,8 @@ use LastDragon_ru\LaraASP\GraphQL\Builder\Contracts\Scout\FieldResolver;
 use LastDragon_ru\LaraASP\GraphQL\Builder\Property;
 use LastDragon_ru\LaraASP\GraphQL\Exceptions\NotImplemented;
 use LastDragon_ru\LaraASP\GraphQL\SortBy\Contracts\Sorter;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Direction;
-use LastDragon_ru\LaraASP\GraphQL\SortBy\Nulls;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Enums\Direction;
+use LastDragon_ru\LaraASP\GraphQL\SortBy\Enums\Nulls;
 use Override;
 
 /**
